# perf: optimize Vector4 Operations (max 4.8x speedup)