**Description**

This PR optimizes Vector 4 Operations by 
- Removing redundant prototype reassignments during each instantiation 
- cache vector component values
- reduce repeated property lookups
- optimizing math operations

The benchmark results for the optimization shows significant improvements : 

| Method                 | Original (ops/sec) | Optimized (ops/sec) | Speedup (x) |
|------------------------|--------------------|----------------------|-------------|
| add                    | 79,211,505         | 191,722,599          | 2.42        |
| addVectors             | 59,301,907         | 189,711,666          | 3.20        |
| addScaled              | 53,866,382         | 188,051,026          | 3.49        |
| applyMatrix4           | 113,162,966        | 184,389,906          | 1.63        |
| clamp                  | 56,666,129         | 188,645,167          | 3.33        |
| lerp                   | 53,043,448         | 185,768,139          | 3.50        |
| setAxisAngleFromQuaternion | 94,430,381    | 185,446,446          | 1.96        |
| componentAccess        | 97,533,182         | 117,665,291          | 1.21        |
| lerpVectors            | 38,556,768         | 185,277,642          | 4.80        |
| equals                 | 84,067,651         | 187,238,687          | 2.23        |