@Mugen87 2025-03-23
How did you measure the performance? Do you mind sharing a live example so it's possible to measure on different devices.

In general, performance metrics are hard to track in the web context since browser/device combinations can produces completely different results for the same test. 

TBH, I'm not in favor of such code changes like in the math methods since it's not obvious to me why a JS engine would execute the new code faster. As a developer, I would prefer to not be forced into such a coding pattern on that level just to improve performance.