@gkjohnson 2025-03-23
I'm also a bit suspicious of some of these benchmark values and am curious as to how they were taken. The `equals` and `add` functions, for example, do not contain any meaningful changes that would seem to impact performance. The same amount of property lookups are done in each, for example, but the table shows over a 2x improvement.

If there are meaningful performance improvements due to something like avoiding redundant local properties reads by assigning to a local variable first then I think it would be best to change one specific function and demonstrate the improvement with a reproduce-able benchmark.