@mattverse 2025-03-23
The benchmark tests were designed and executed as following: 

```
import Benchmark from 'benchmark';
import { Vector4 as OriginalVector4 } from '../../../src/math/Vector4.js';
import { Vector4 as OptimizedVector4 } from '../../../src/math/Vector4Optimized.js';

const MATRIX = {
	elements: [
		2, 3, 5, 7,
		11, 13, 17, 19,
		23, 29, 31, 37,
		41, 43, 47, 53
	]
};

const suite = new Benchmark.Suite();

// Benchmark configuration
const config = {
	minSamples: 100,
	initCount: 10,
	delay: 0.02
};

suite
/* add() */
	.add( 'Original add', () => {

		const a = new OriginalVector4( 1, 2, 3, 4 );
		const b = new OriginalVector4( 5, 6, 7, 8 );
		a.add( b );

	}, config )
	.add( 'Optimized add', () => {

		const a = new OptimizedVector4( 1, 2, 3, 4 );
		const b = new OptimizedVector4( 5, 6, 7, 8 );
		a.add( b );

	}, config )

```

I was also surprised to see how much of perf improvement this did but depending on how we want to take code patterns etc. I'm happy to either close the PR or reduce the scope of this PR