@Mugen87 2025-03-23
I'm afraid without in-depth testing with different browsers/devices such PRs are not acceptable.